move most libs to core in preparation for flywheel pkg
