+++ baker-endpoint/package.json
@@ -27,7 +27,8 @@
   },
   "dependencies": {
     "@flashbake/core": "file:../core",
-    "@taquito/rpc": "^16.1.2",
+    "@taquito/rpc": "^17.0.0",
+    "@taquito/taquito": "^17.0.0",
     "@taquito/utils": "^16.1.2",
     "@types/yargs": "^17.0.10",
     "body-parser": "^1.19.1",
