+++ core/package.json
@@ -21,9 +21,11 @@
     "typescript": "^4.4.4"
   },
   "dependencies": {
-    "@taquito/local-forging": "^16.1.2",
-    "@taquito/rpc": "^16.1.2",
+    "@taquito/local-forging": "^17.0.0",
+    "@taquito/rpc": "^17.0.0",
+    "@taquito/taquito": "^17.0.0",
     "bs58check-ts": "^0.0.7",
-    "lodash": "^4.17.21"
+    "lodash": "^4.17.21",
+    "p-limit": "^3.1.0"
   }
 }
