+++ core/src/implementations/caching-baking-rights-service.ts
@@ -1,7 +1,7 @@
-import TtlWindowMonitor, { TtlWindowObserver } from "../../interfaces/ttl-window-monitor";
-import BakingRightsService, { BakingAssignment, BakingMap } from "../../interfaces/baking-rights-service";
-import { BlockMonitor, BlockNotification } from '@flashbake/core';
-import { RegistryService } from '../../interfaces/registry-service';
+import TtlWindowMonitor, { TtlWindowObserver } from "../interfaces/ttl-window-monitor";
+import BakingRightsService, { BakingAssignment, BakingMap } from "../interfaces/baking-rights-service";
+import BlockMonitor, { BlockNotification } from '../interfaces/block-monitor';
+import { RegistryService } from '../interfaces/registry-service';
 import * as http from "http";
 import pLimit from "p-limit";
