+++ core/src/index.ts
@@ -5,12 +5,22 @@
 // Types
 export { Bundle } from './types/bundle'
 export { Address, TransactionHash } from './types/core'
+export { RegistryValue } from './types/registry-value'
 export { TezosParsedOperation } from './types/tezos-parsed-operation'
 export { default as BlockMonitor, BlockObserver, BlockNotification } from './interfaces/block-monitor'
 
 // Classes
 export { default as BundleUtils } from './bundle-utils'
+export { RegistryService } from './interfaces/registry-service'
+export { default as BakingRightsService, BakingAssignment } from './interfaces/baking-rights-service'
+export { default as TtlWindowMonitor, TtlWindowObserver } from './interfaces/ttl-window-monitor'
 export { default as TezosOperationUtils } from './operation-utils'
 export { default as Network } from './network'
 export { default as RpcBlockMonitor } from './implementations/rpc-block-monitor'
 export { default as ConstantsUtil } from './implementations/rpc-constants'
+export { default as CachingBakingRightsService } from './implementations/caching-baking-rights-service'
+export { default as InMemoryRegistryService } from './implementations/in-memory-registry-service'
+export { default as OnChainRegistryService } from './implementations/on-chain-registry-service'
+export { default as GenericTtlWindowMonitor } from './implementations/generic-ttl-window-monitor'
+export { default as RpcTtlWindowMonitor } from './implementations/rpc-ttl-window-monitor'
+export { default as TaquitoRpcService } from './implementations/taquito-rpc-service'
