+++ core/src/types/registry-value.ts
@@ -1,0 +1,6 @@
+import { Address } from "../types/core";
+
+export type RegistryValue = {
+  // Participating baker base URL
+  endpointUrl: string;
+}
