+++ relay/package.json
@@ -26,8 +26,9 @@
   },
   "dependencies": {
     "@flashbake/core": "file:../core",
-    "@taquito/rpc": "^16.1.2",
-    "@taquito/taquito": "^16.1.2",
+    "@taquito/local-forging": "^17.0.0",
+    "@taquito/rpc": "^17.0.0",
+    "@taquito/taquito": "^17.0.0",
     "@taquito/utils": "^16.1.2",
     "@types/yargs": "^17.0.10",
     "axios": "^0.24.0",
@@ -35,7 +36,6 @@
     "express": "^4.17.2",
     "http-proxy-middleware": "^2.0.1",
     "lodash": "^4.17.21",
-    "p-limit": "^3.1.0",
     "prom-client": "^14.0.1",
     "trace-unhandled": "^2.0.1",
     "yargs": "^17.5.1"
