+++ relay/src/index.ts
@@ -2,21 +2,6 @@
  * Export all components.
  */
 
-// Types
-export { RegistryValue } from './types/registry-value'
-
-// Interfaces
-export { default as IndexerService } from './interfaces/indexer-service'
-export { RegistryService } from './interfaces/registry-service'
-export { default as RpcService } from './interfaces/rpc-service'
-export { default as BakingRightsService, BakingAssignment } from './interfaces/baking-rights-service'
-export { default as TtlWindowMonitor, TtlWindowObserver } from './interfaces/ttl-window-monitor'
 
 // Classes
-export { default as InMemoryRegistryService } from './implementations/in-memory/in-memory-registry-service'
-export { default as OnChainRegistryService } from './implementations/taquito/on-chain-registry-service'
-export { default as TaquitoRpcService } from './implementations/taquito/taquito-rpc-service'
-export { default as GenericTtlWindowMonitor } from './implementations/in-memory/generic-ttl-window-monitor'
-export { default as RpcTtlWindowMonitor } from './implementations/rpc/rpc-ttl-window-monitor'
-export { default as CachingBakingRightsService } from './implementations/in-memory/caching-baking-rights-service'
 export { default as HttpRelay } from './http-relay'
