--- relay/src/types/registry-value.ts
@@ -1,6 +1,0 @@
-import { Address } from "@flashbake/core"
-
-export type RegistryValue = {
-    // Participating baker base URL
-    endpointUrl: string;
-}
